errorClass element when using row / col
When placing form elements in grid, the error class is set on the outer form-control div. If you're form elements span over more columns inside the form-group, the entire form-group (all elements) is highlighted with error.

Example:

<div class="row form-group">
	<div class="col-xs-6">
		<label class="control-label" for="TimeStart">TimeStart</label>
		<input class="form-control" id="TimeStart" name="TimeStart" type="text" value="" />
	</div>
	<div class="col-xs-6">
		<label class="control-label" for="Hours">Hours</label>
		<input class="form-control" id="Hours" name="Hours" type="text" value="" />
	</div>
</div>

This checks if the form-control element also contains a "row" class. If form elements is placed inside a row element, then it will add the error className to the sub elements which has a className starting with "col-*".